Fix bug when the cart_id is not present in cart_item

+++ controllers/cartController.js
@@ -1,4 +1,5 @@
 const { getAllCartItemsSql, createCartItemsSql, getSingleCartItemSql, updateCartSql, deleteCartItemSql } = require("../model/cartModel");
+const db = require("../db/connect");
 
 
 
@@ -52,8 +53,16 @@ const deleteCartItem = async (req,res) => {
     try {
         const {id} = req.params;
         const {product_id} = req.body;
+        let sql = `SELECT * FROM cart_item  WHERE product_id = '${product_id}' and cart_id ='${id}'`
+        const [cartItemExists,_] = await db.execute(sql);
+        if(cartItemExists.length === 0){
+            return res.status(404).send({
+                status : 404,
+                msg : `cart with cart_id = ${id} and product_id = ${product_id} not found`
+            })
+        }
         const cart = await deleteCartItemSql(id,product_id);
-        res.send(cart);
+        res.send({msg : `cart item with cart_id = ${id} and product_id = ${product_id} deleted`,cart : cartItemExists});
     } catch (error) {
         console.log(error);
         res.status(404).send({msg : error});
